+++ towerhttp/respond_error.go
@@ -22,7 +22,7 @@ func (r Responder) RespondError(ctx context.Context, rw http.ResponseWriter, err
 	body := r.errorTransformer.ErrorBodyTransform(ctx, err)
 	b, errIO := opt.encoder.Encode(body)
 	if errIO != nil {
-		_ = r.tower.Wrap(errIO).Caller(tower.GetCaller(3)).Log(ctx)
+		_ = r.tower().Wrap(errIO).Caller(tower.GetCaller(3)).Log(ctx)
 		return
 	}
 	contentType := opt.encoder.ContentType()
@@ -31,7 +31,7 @@ func (r Responder) RespondError(ctx context.Context, rw http.ResponseWriter, err
 	}
 	compressed, errIO := opt.compressor.Compress(b)
 	if errIO != nil {
-		_ = r.tower.Wrap(errIO).Caller(tower.GetCaller(3)).Log(ctx)
+		_ = r.tower().Wrap(errIO).Caller(tower.GetCaller(3)).Log(ctx)
 		return
 	}
 	contentEncoding := opt.compressor.ContentEncoding()
@@ -42,6 +42,6 @@ func (r Responder) RespondError(ctx context.Context, rw http.ResponseWriter, err
 	rw.WriteHeader(opt.statusCode)
 	_, errIO = rw.Write(compressed)
 	if errIO != nil {
-		_ = r.tower.Wrap(errIO).Caller(tower.GetCaller(3)).Log(ctx)
+		_ = r.tower().Wrap(errIO).Caller(tower.GetCaller(3)).Log(ctx)
 	}
 }